+++ nacc_attribute_deriver/attributes/nacc/modules/uds/__init__.py
@@ -2,5 +2,7 @@
 from .form_a1 import *
 from .form_a2 import *
 from .form_a3 import *
+from .form_a4 import *
+from .form_a5 import *
 from .form_d1 import *
 from .form_header import *